Generated by Spark: There is a new "TMNT: Shredder and Krang" hero pack coming associated with teenage mutant ninja turtles.  Krang as a hero has starting health 16, melee, movement 1.  His special ability is called "DOOOOM!" and his special ability says "Krang has 3 doomsday machines. Start with one machine active. After you roll the die of epic destruction, you can flip an active machine to reroll the die. Add +1 to your move value for each active machine."
add krang as a hero and his set should be able to be added in collections.

Shredder is melee with starting health 15. His sidekick(s) are Bebop & Rocksteady (represented as a single sidekick), who are melee with starting health of 7.  The movement value is 3 for shredder and his minion.  The special ability is called "Master of the clan" and reads "At the start of your turn, deploy a Foot soldier to a path adjacent to a friendly fighter. You may attack opposing fighters adjacent to Foot soldiers. If an opponent boosts their maneuver, they may remove any Foot soldiers their hero moves through."

Add these two heroes